refactor: rdbフォルダをmysqlにリネーム
