+++ internal/infrastructure/mysql/transaction.go
@@ -1,4 +1,4 @@
-package rdb
+package mysql
 
 import (
 	"context"
